+++ testsuite/tests/properties/entity_bind/test.py
@@ -4,7 +4,7 @@ Test that Bind works when binding from entities.
 
 from __future__ import absolute_import, division, print_function
 
-from langkit.dsl import ASTNode, LogicVarType, LongType, UserField
+from langkit.dsl import ASTNode, LogicVarType, IntegerType, UserField
 from langkit.expressions import AbstractProperty, Let, Property, Self, Bind
 from langkit.parsers import Grammar
 
@@ -13,7 +13,7 @@ from utils import build_and_run
 
 
 class FooNode(ASTNode):
-    prop = AbstractProperty(runtime_check=True, type=LongType, public=True)
+    prop = AbstractProperty(runtime_check=True, type=IntegerType, public=True)
 
 
 class Literal(FooNode):
