+++ testsuite/tests/properties/invalid_prop_args_2/test.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from langkit.dsl import ASTNode, Field, LongType, SymbolType
+from langkit.dsl import ASTNode, Field, IntegerType, SymbolType
 from langkit.expressions import Property, langkit_property, AbstractKind
 from langkit.parsers import Grammar
 
@@ -14,8 +14,8 @@ class FooNode(ASTNode):
 class BarCode(FooNode):
     a = Field()
 
-    @langkit_property(kind=AbstractKind.abstract, return_type=LongType)
-    def base_prop(x=LongType):
+    @langkit_property(kind=AbstractKind.abstract, return_type=IntegerType)
+    def base_prop(x=IntegerType):
         pass
